refactor(actions and events): :truck: move event handlers to root level, move actions to their parent folder
