+++ src/app/event-handlers/provider-application-event-handlers.ts
@@ -1,4 +1,4 @@
-import { provideApplicationEventHandler } from '../event-handler/provide-application-event-handler.function';
+import { provideApplicationEventHandler } from '../event-hub/event-handler/provide-application-event-handler.function';
 import { LanguageChangedEventHandler } from './language-changed-event-handler.service';
 
 export const APPLICATION_EVENT_HANDLERS = [
