+++ src/app/internalization/lang-store/lang-store.service.ts
@@ -1,4 +1,3 @@
-import { languageChange$ } from '@actions/lang.actions';
 import { effect, inject, Injectable } from '@angular/core';
 import { toSignal } from '@angular/core/rxjs-interop';
 import { isLang } from '@internalization/is-lang.function';
@@ -16,6 +15,7 @@ import { filter, merge, of, switchMap, tap } from 'rxjs';
 import { APPLICATION_BUS } from 'src/app/event-hub/event-bus/event-bus.token';
 import { LanguageChangedEvent } from 'src/app/events/language-change.event';
 import { initialLanguageState } from './lang-state.type';
+import { languageChange$ } from './lang.actions';
 import { languageAdapter } from './language.adapter';
 
 @Injectable({
