+++ tsconfig.json
@@ -26,7 +26,6 @@
     "paths": {
       "@shared/*": ["src/app/shared/*"],
       "@theme/*": ["src/app/theme/*"],
-      "@actions/*": ["src/app/actions/*"],
       "@store/*": ["src/app/store/*"],
       "@about/*": ["src/app/about/*"],
       "@welcome/*": ["src/app/welcome/*"],
